% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_diff.R
\name{my_diff}
\alias{my_diff}
\title{Calculate a difference}
\usage{
my_diff(a, b)
}
\arguments{
\item{a}{numeric value}

\item{b}{numeric value}
}
\value{
numeric value; the difference between the input variables
}
\description{
Calculate a difference
}
\examples{
my_diff(5, 3)
}
