% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{my_sum}
\alias{my_sum}
\title{Sum up values}
\usage{
my_sum(a, b)
}
\arguments{
\item{a}{A numeric value.}

\item{b}{A numeric value.}
}
\value{
The sum of the two values.
}
\description{
This function sums up two values.
}
\examples{
my_sum(2, 3) # returns 5
my_sum(-1, 1) # returns 0
}
\seealso{
sum
}
